requirements.txt
fastapi~=0.114.2
uvicorn~=0.30.6
google-cloud-storage
alembic~=1.13.2
sqlalchemy~=2.0.34
vertexai~=1.66.0
gradio~=4.44.0
google-cloud-aiplatform~=1.66.0
pip~=23.2.1
pillow~=10.4.0
Mako~=1.3.5
psycopg2
